# Optimize TableColumn component by caching relationship checks