## Summary
<!-- Briefly describe the changes and the purpose of the PR. -->
Optimize TableColumn component by caching relationship checks.
Introduced useMemo to cache `isSource` and `targetCardinality` computations.



## Experiment

Triggered `handleMouseLeaveNode` manually on the accounts table node to measure performance impact.

<details><summary>Demo</summary>

https://github.com/user-attachments/assets/9cde8ec3-49ca-455a-afe5-a03f751fba1d


</details>





#### 1.  `ERDContentInner`

before: ~660 ms
after: ~170 ms


|  | before 90fae17723f7 | after f843a99746a2c37  |
|--------|--------|--------|
|  | <img width="1216" alt="main1-sk" src="https://github.com/user-attachments/assets/870f0f61-6ae8-46a4-b5c6-9565682551b7" />  | <img width="1207" alt="hoshino1-sk" src="https://github.com/user-attachments/assets/f5c0e444-e7b9-46fd-94bc-d5f636d58a53" /> |





#### 2. `TableColumn`: `account_aliases` table > `id` column

before: ~0.6 ms
after: ~0.1 ms


|  | before 90fae17723f7 | after  f843a99746a2c37 |
|--------|--------|--------|
|  | <img width="536" alt="main2-sk" src="https://github.com/user-attachments/assets/911e64c1-e7b8-4337-a439-eccdd2b104b4" /> |  <img width="539" alt="hoshino2-sk" src="https://github.com/user-attachments/assets/0e3f0e50-1ee5-40f2-9de1-1d54f06742e6" /> | 




## Testing
<!-- Briefly describe the testing steps or results. -->

N/A

## Other Information
<!-- Add any other relevant information for the reviewer. -->

N/A